// shared-schemas/src/types.d.ts
export interface Point2D {
    x_mm: number;
    y_mm: number;
}
export interface Point3D {
    x_mm: number;
    y_mm: number;
    z_mm: number;
}
export interface Vector2D {
    x: number;
    y: number;
}
// ...
export interface WallSegment2D {
    id: string;
    start: Point2D;
    end: Point2D;
    length_mm: number;
}
// ...
export interface PlacedItem {
    id: string;
    wall_id: string;
    offset_mm: number;
    width_mm: number;
}